style: :truck: Moved add trip form to shared
