+++ src/app/components/trip/trip.component.ts
@@ -7,7 +7,7 @@ import {
   matCalendarMonth,
 } from '@ng-icons/material-icons/baseline';
 import { NgIconComponent, provideIcons } from '@ng-icons/core';
-import { AddTripFormComponent } from '../add-trip-form/add-trip-form.component';
+import { AddTripFormComponent } from '../shared/add-trip-form/add-trip-form.component';
 import { Store } from '@ngrx/store';
 import { CommonModule } from '@angular/common';
 import { TripState } from '../../store/state';
